DM-13112: Connect portal to user file workspace when available (TODO: support for dynamic updates to file list on client side, when folder node is clicked)
